refactor: componentized home page code
