+++ pages/index.tsx
@@ -1,17 +1,12 @@
 'use client';
-import { useState, useEffect, ChangeEvent, KeyboardEvent } from 'react';
+import { useState, useEffect } from 'react';
 import dynamic from 'next/dynamic';
-import Image from 'next/image';
-import {
-    createParser,
-    ParsedEvent,
-    ReconnectInterval,
-} from 'eventsource-parser';
 
+import { IChannel } from '@yomo/presence';
 import { presenceConnect } from '../helper/presenceConnect';
 import { Header } from '../components/ChatContainer/Header';
 import { MessageContainer } from '../components/ChatContainer/MessageContainer';
-import { IChannel } from '@yomo/presence';
+import { UserInput } from '../components/UserInput';
 const UserCursor = dynamic(
     () => import('../components/UserCursor').then((mod) => mod.UserCursor),
     { ssr: false }
@@ -24,7 +19,7 @@ const currentConnectId = (
 const defaultMessage: Message = {
     role: 'assistant',
     state: 'deltaStart',
-    content: 'Welcome to CollabGPT!',
+    content: 'Welcome to Allegro CollabGPT!',
     avatar: '',
 };
 
@@ -43,24 +38,25 @@ export default function Home() {
     const [messages, setMessages] = useState<Message[]>([defaultMessage]);
     const [onlineUsers, setOnlineUsers] = useState<UserInfo[]>([]);
 
-    const [userInput, setUserInput] = useState<string>('');
     const [loadingState, setLoadingState] = useState<boolean>(false);
 
-    const appendMessages = (deltaMessage: Message) => {
-        channel?.broadcast('chatInfo', { ...deltaMessage });
+    const userInputSetLoadingState = (state: boolean) => {
+        setLoadingState(state);
+    };
+    const userInputAppendMessage = (deltaMessage: Message) => {
         setMessages((messages) => [...messages, deltaMessage]);
     };
-    const modifyLastUserinput = (inputMessage: string) => {
+    const userInputSetInputMessage = (inputContent: string) => {
         setMessages((messages) => {
             const lastMessage = messages[messages.length - 1];
             const updatedMessage = {
                 ...lastMessage,
-                content: inputMessage,
+                content: inputContent,
             };
             return [...messages.slice(0, -1), updatedMessage];
         });
     };
-    const modifyLastMessages = (deltaMessage: Message) => {
+    const userInputSetReceivedMessage = (deltaMessage: Message) => {
         setMessages((messages) => {
             const lastMessage = messages[messages.length - 1];
             const updatedMessage = {
@@ -69,95 +65,6 @@ export default function Home() {
             };
             return [...messages.slice(0, -1), updatedMessage];
         });
-    };
-
-    const broadcastLoadingState = (state: boolean) => {
-        channel?.broadcast('loadingState', { isLoading: state });
-        setLoadingState(state);
-    };
-
-    const broadcastMessages = (deltaMessage: Message) => {
-        channel?.broadcast('chatInfo', { ...deltaMessage });
-        modifyLastMessages(deltaMessage);
-    };
-
-    const handleHttpRequestError = () => {
-        broadcastLoadingState(false);
-        broadcastMessages({
-            state: 'deltaStart',
-            role: 'assistant',
-            content: 'Error',
-            avatar: '',
-        });
-        setUserInput('');
-    };
-
-    const broadcastInputMessage = (event: ChangeEvent<HTMLTextAreaElement>) => {
-        setUserInput(() => event.target.value);
-        modifyLastUserinput(event.target.value);
-        channel?.broadcast('chatInfo', {
-            state: 'input',
-            role: 'user',
-            content: event.target.value,
-        });
-    };
-
-    const submitInput = async () => {
-        if (!userInput) return;
-
-        broadcastLoadingState(true);
-
-        const response = await fetch('/api/chat', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ message: userInput }),
-        });
-        if (!response.ok) {
-            handleHttpRequestError();
-            return;
-        }
-        const data = response.body?.getReader();
-        if (!data) {
-            handleHttpRequestError();
-            return;
-        }
-
-        appendMessages({
-            state: 'deltaStart',
-            role: 'assistant',
-            content: '',
-            avatar: '',
-        });
-
-        const decoder = new TextDecoder();
-        const parser = createParser(
-            (event: ParsedEvent | ReconnectInterval) => {
-                if (event.type === 'event') {
-                    const data = event.data;
-                    if (data === '[DONE]') return;
-                    const json = JSON.parse(data);
-                    if (json.choices[0].finish_reason === 'stop') return;
-                    const chunkValue = json.choices[0].delta.content;
-                    broadcastMessages({
-                        state: 'delta',
-                        role: 'assistant',
-                        content: chunkValue,
-                        avatar: '',
-                    });
-                }
-            }
-        );
-
-        while (true) {
-            const { value, done } = await data?.read();
-            if (done) break;
-            parser.feed(decoder.decode(value));
-        }
-
-        setUserInput('');
-        broadcastLoadingState(false);
     };
 
     useEffect(() => {
@@ -184,52 +91,15 @@ export default function Home() {
             />
             <Header onlineUsers={onlineUsers} />
             <MessageContainer messages={messages} loading={loadingState} />
-
-            <div className="relative mx-auto my-6 flex w-3/5 flex-nowrap items-center justify-center">
-                <textarea
-                    className={`min-h-12 w-full whitespace-nowrap rounded-lg border border-[#34323E] bg-[#171820] p-2 focus:outline-none focus:border-[${defaultUserInfo.color}]`}
-                    disabled={loadingState}
-                    style={{ resize: 'none' }}
-                    placeholder="Type your query..."
-                    value={userInput}
-                    rows={1}
-                    onChange={(event) => broadcastInputMessage(event)}
-                    onFocus={() => {
-                        appendMessages({
-                            state: 'inputStart',
-                            role: 'user',
-                            content: `user${currentConnectId} is typing...`,
-                            avatar: defaultUserInfo.avatar,
-                        });
-                        channel?.broadcast('loadingState', { isLoading: true });
-                    }}
-                    onBlur={() => {
-                        channel?.broadcast('loadingState', {
-                            isLoading: false,
-                        });
-                    }}
-                    onKeyDown={(event: KeyboardEvent<HTMLTextAreaElement>) => {
-                        if (event.key === 'Enter') submitInput();
-                    }}
-                />
-                <button
-                    onClick={submitInput}
-                    disabled={loadingState}
-                    className="absolute right-2"
-                >
-                    <Image
-                        src={'/send-arrow.svg'}
-                        alt="send arrow"
-                        width={24}
-                        height={24}
-                        className={`${
-                            loadingState
-                                ? 'hover:cursor-not-allowed'
-                                : 'hover:cursor-pointer'
-                        } hover:opacity-80`}
-                    />
-                </button>
-            </div>
+            <UserInput
+                channel={channel}
+                defaultUserInfo={defaultUserInfo}
+                loadingState={loadingState}
+                userInputSetLoadingState={userInputSetLoadingState}
+                userInputAppendMessage={userInputAppendMessage}
+                userInputSetInputMessage={userInputSetInputMessage}
+                userInputSetReceivedMessage={userInputSetReceivedMessage}
+            />
         </main>
     );
 }
